==> analysis/6_results_spatial.Rmd <==
---
title: "Results"
output: html_notebook
author: Guy Lomax
date: 2024-01-12
editor_options: 
  chunk_output_type: console
---

This notebook compares the results of the RPI methodology with those of RUE,
RESTREND and LGS at identifying persistent differences in multi-annual mean
productivity.

```{r setup}

library(tidyverse)
library(terra)
library(sf)
library(here)
library(tmap)

```


```{r load}

# Country polygons
ke_tz <- st_read(here("data", "raw", "vector", "kenya_tanzania.geojson"))

# Actual GPP mean
rpi_paths <- Sys.glob(here("data", "processed", "raster", "rpi", "rf_rpi_100_20*.tif"))

potential_gpp_mean <- rpi_paths %>%
  map(function(path) rast(path) %>% subset("quantile_pred")) %>%
  rast() %>%
  mean()

actual_gpp_mean <- rpi_paths %>%
  map(function(path) rast(path) %>% subset("GPP")) %>%
  rast() %>%
  mean()

predicted_gpp_rpi <- rpi_paths %>%
  map(function(path) rast(path) %>% subset("mean_pred")) %>%
  rast() %>%
  mean()

# RPI mean and trend
rpi_mean <- rast(here("data", "processed", "raster", "rpi", "rpi_mean.tif"))

# RUE mean and trend
rue_mean <- rast(here("data", "processed", "raster", "rue", "rue_mean.tif"))

# LGS mean

lgs_k <- c(100, 200, 1000)
lgs_paths <- Sys.glob(here("data", "processed", "raster", "lgs", "lgs_subset_*.tif"))

lgs_layers <- lgs_paths %>%
  map(rast)
names(lgs_layers) <- lgs_k

lgs_100 <- lgs_layers[["100"]]
lgs_200 <- lgs_layers[["200"]]
lgs_1000 <- lgs_layers[["1000"]]

# Key covariates for comparison/exploration of results
dynamic_covariate_paths <- Sys.glob(here("data", "raw", "raster", "covariateMaps", "dynamicVars*.tif"))

years <- 2000:2018
ppt_all <- dynamic_covariate_paths %>%
  map(function(path) rast(path) %>% subset("precipitation")) %>%
  rast()
t_mean_all <- dynamic_covariate_paths %>%
  map(function(path) rast(path) %>% subset("tMean")) %>%
  rast()
tree_cover <- rast(here("data", "raw", "raster", "covariateMaps", "staticVars.tif")) %>%
  subset("wriTCFrac") %>%
  crop(ppt_all)

names(ppt_all) <- paste0("ppt_", years)
names(t_mean_all) <- paste0("tMean_", years)
names(tree_cover) <- "tree_cover"

ppt_mean <- mean(ppt_all)
t_mean <- mean(t_mean_all)
covariates <- c(ppt_mean, t_mean, tree_cover)
names(covariates) <- c("ppt_mean", "t_mean", "tree_cover")


```


We can estimate the usefulness of each method in comparing spatial patterns
by quantifying how much of the spatial variation in GPP it is able to explain
using environmental variation. For RUE this is extremely crude, but for the
others it is more meaningful.

```{r spatial_variance_explained}

mean_gpp_value <- mean(values(actual_gpp_mean), na.rm = T)

tss_gpp <- (actual_gpp_mean - mean_gpp_value) ^ 2 %>%
  values() %>%
  sum(na.rm = TRUE)

# RUE

# median_rue <- median(values(rue_mean), na.rm = T)
mean_rue <- mean(values(rue_mean), na.rm = T)

predicted_gpp_rue <- mean_rue * mean(ppt_all)

abs_error_rue <- abs(actual_gpp_mean - predicted_gpp_rue)

mae_rue <- mean(values(abs_error_rue), na.rm = TRUE)

rss_gpp_rue <- (actual_gpp_mean - predicted_gpp_rue) ^ 2 %>%
  values() %>%
  sum(na.rm = TRUE)

rsq_rue <- 1 - (rss_gpp_rue / tss_gpp)

# LGS

abs_error_lgs <- abs(actual_gpp_mean - lgs_1000$mean)

mae_lgs <- mean(values(abs_error_lgs), na.rm = T)

rss_gpp_lgs <- (actual_gpp_mean - lgs_1000$mean) ^ 2 %>%
  values() %>%
  sum(na.rm = T)

rsq_lgs <- 1 - (rss_gpp_lgs / tss_gpp)

# RPI

mae_rpi <- mean(values(abs(actual_gpp_mean - predicted_gpp_rpi)), na.rm = T)

rss_gpp_rpi <- (actual_gpp_mean - predicted_gpp_rpi) ^ 2 %>%
  values() %>%
  sum(na.rm = TRUE)

rsq_rpi <- 1 - (rss_gpp_rpi / tss_gpp)

```



Plot comparative maps of the results from the different methods (Figure 4).

```{r spatial_comparison_maps}

# Create new bbox to allow space for legend

map_bbox <- st_bbox(ke_tz)

xrange <- map_bbox$xmax - map_bbox$xmin
yrange <- map_bbox$ymax - map_bbox$ymin

map_bbox[1] <- map_bbox[1] - 0.25 * xrange

# Basemap elements
fill <- tm_shape(ke_tz, bbox = map_bbox) +
  tm_fill("grey95")

borders <- tm_shape(ke_tz) +
  tm_borders() +
  tm_layout(frame = FALSE)

zoom_box <- st_bbox(c(xmin = 38, xmax = 40, ymin = -0.5, ymax = 1.5)) %>%
  st_as_sfc(crs = "EPSG:4326") %>%
  st_as_sf()

zoom_borders <- tm_shape(zoom_box) +
  tm_borders(col = "red", lwd = 1.5)

# Raw maps of mean values
rpi_mean_map <- fill +
  tm_shape(rpi_mean) +
  tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(0.3, 1),
      values = "-viridis",
      outliers.trunc = c(T,T)
    ),
    col.legend = tm_legend(
      title = "",
      reverse = TRUE,
      frame = FALSE,
      position = c("left", "top")
    )
  ) +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  tm_title("a) RPI") +
  borders +
  zoom_borders

rue_mean_map <- fill +
  tm_shape(rue_mean) +
  tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(1, 7),
      values = "-viridis",
      outliers.trunc = c(T,T)
    ),
    col.legend = tm_legend(
      title = "",
      reverse = TRUE,
      frame = FALSE,
      position = c("left", "top")
    )
  ) +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  tm_title("b) RUE") +
  borders +
  zoom_borders

lgs_100_mean_map <- fill +
  tm_shape(lgs_100$lgs_ratio) +
  tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(0.25, 1.25),
      values = "-viridis",
      outliers.trunc = c(T,T)
    ),
    col.legend = tm_legend(
      title = "",
      reverse = TRUE,
      frame = FALSE,
      position = c("left", "top")
    )
  ) +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  tm_title("c) LGS, k = 100") +
  borders +
  zoom_borders

lgs_200_mean_map <- fill +
  tm_shape(lgs_200$lgs_ratio) +
  tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(0.25, 1.25),
      values = "-viridis",
      outliers.trunc = c(T,T)
    ),
    col.legend = tm_legend(
      title = "",
      reverse = TRUE,
      frame = FALSE,
      position = c("left", "top")
    )
  ) +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  tm_title("d) LGS, k = 200") +
  borders +
  zoom_borders

lgs_1000_mean_map <- fill +
  tm_shape(lgs_1000$lgs_ratio) +
  tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(0.25, 1.25),
      values = "-viridis",
      outliers.trunc = c(T,T)
    ),
    col.legend = tm_legend(
      title = "",
      reverse = TRUE,
      frame = FALSE,
      position = c("left", "top")
    )
  ) +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  tm_title("d) LGS, k = 1000") +
  borders +
  zoom_borders

# Save maps
tmap_save(rue_mean_map,
          here("results", "figures", "rue_mean.png"),
          width = 12, height = 12, units = "cm", dpi = 250)

tmap_save(rpi_mean_map,
          here("results", "figures", "rpi_mean.png"),
          width = 12, height = 12, units = "cm", dpi = 250)

tmap_save(lgs_100_mean_map,
          here("results", "figures", "lgs_100_mean.png"),
          width = 12, height = 12, units = "cm", dpi = 250)

tmap_save(lgs_200_mean_map,
          here("results", "figures", "lgs_200_mean.png"),
          width = 12, height = 12, units = "cm", dpi = 250)

tmap_save(lgs_1000_mean_map,
          here("results", "figures", "lgs_1000_mean.png"),
          width = 12, height = 12, units = "cm", dpi = 250)

## The same maps zoomed in to a specific region

zoom_raster <- c(rue_mean, rpi_mean, lgs_100$lgs_ratio, lgs_1000$lgs_ratio)
names(zoom_raster) <- c("rue", "rpi", "lgs_100", "lgs_1000")

zoom_raster <- crop(zoom_raster, zoom_box)

zoom_rue <- tm_shape(zoom_raster$rue) + 
  tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(1, 7),
      values = "-viridis",
      outliers.trunc = c(T,T)
    ),
    col.legend = tm_legend(
      show = FALSE
    )
  ) +
  tm_shape(zoom_box) + tm_borders(col = "red", lwd = 2) +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  tm_scalebar()

zoom_rpi <- tm_shape(zoom_raster$rpi) + 
  tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(0.3, 1),
      values = "-viridis",
      outliers.trunc = c(T,T)
    ),
    col.legend = tm_legend(
      show = FALSE
    )
  ) +
  tm_shape(zoom_box) + tm_borders(col = "red", lwd = 2) +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  tm_scalebar()

zoom_lgs_100 <- tm_shape(zoom_raster$lgs_100) + 
  tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(0.25, 1.25),
      values = "-viridis",
      outliers.trunc = c(T,T)
    ),
    col.legend = tm_legend(
      show = FALSE
    )
  ) +
  tm_shape(zoom_box) + tm_borders(col = "red", lwd = 2) +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  tm_scalebar()

zoom_lgs_1000 <- tm_shape(zoom_raster$lgs_1000) + 
  tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(0.25, 1.25),
      values = "-viridis",
      outliers.trunc = c(T,T)
    ),
    col.legend = tm_legend(
      show = FALSE
    )
  ) +
  tm_shape(zoom_box) + tm_borders(col = "red", lwd = 2) +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  tm_scalebar()

tmap_save(zoom_rpi,
          here("results", "figures", "zoom_rpi.png"),
          width = 12, height = 12, units = "cm", dpi = 250)

tmap_save(zoom_rue,
          here("results", "figures", "zoom_rue.png"),
          width = 12, height = 12, units = "cm", dpi = 250)

tmap_save(zoom_lgs_100,
          here("results", "figures", "zoom_lgs_100.png"),
          width = 12, height = 12, units = "cm", dpi = 250)

tmap_save(zoom_lgs_1000,
          here("results", "figures", "zoom_lgs_1000.png"),
          width = 12, height = 12, units = "cm", dpi = 250)


```


Plot the estimated potential GPP to get a clearer picture of the model
underlying each method (Supplementary Figure S2):

```{r potential_gpp_maps}

# Create new bbox to allow space for legend

map_bbox <- st_bbox(ke_tz)

xrange <- map_bbox$xmax - map_bbox$xmin

map_bbox[1] <- map_bbox[1] - 0.25 * xrange

# Basemap elements
fill <- tm_shape(ke_tz, bbox = map_bbox) +
  tm_fill("grey95")

borders <- tm_shape(ke_tz) +
  tm_borders() +
  tm_layout(frame = FALSE)

zoom_box <- st_bbox(c(xmin = 38, xmax = 40, ymin = -0.5, ymax = 1.5)) %>%
  st_as_sfc(crs = "EPSG:4326") %>%
  st_as_sf()

zoom_borders <- tm_shape(zoom_box) +
  tm_borders(col = "red", lwd = 1.5)

# Raw maps of mean values
rpi_pgpp_map <- fill +
  tm_shape(potential_gpp_mean) +
  tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(0, 5000),
      values = "yl_gn",
      outliers.trunc = c(T,T)
    ),
    col.legend = tm_legend(
      title = "",
      reverse = TRUE,
      frame = FALSE,
      position = c("left", "top")
    )
  ) +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  tm_title("a) RPI") +
  borders +
  zoom_borders

rue_pgpp_map <- fill +
  tm_shape(mean(rue_pred)) +
  tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(0, 5000),
      values = "yl_gn",
      outliers.trunc = c(T,T)
    ),
    col.legend = tm_legend(
      title = "",
      reverse = TRUE,
      frame = FALSE,
      position = c("left", "top")
    )
  ) +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  tm_title("b) RUE") +
  borders +
  zoom_borders

lgs_100_pgpp_map <- fill +
  tm_shape(lgs_100$potential) +
  tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(0, 5000),
      values = "yl_gn",
      outliers.trunc = c(T,T)
    ),
    col.legend = tm_legend(
      title = "",
      reverse = TRUE,
      frame = FALSE,
      position = c("left", "top")
    )
  ) +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  tm_title("c) LGS, k = 100") +
  borders +
  zoom_borders

lgs_1000_pgpp_map <- fill +
  tm_shape(lgs_1000$potential) +
  tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(0, 5000),
      values = "yl_gn",
      outliers.trunc = c(T,T)
    ),
    col.legend = tm_legend(
      title = "",
      reverse = TRUE,
      frame = FALSE,
      position = c("left", "top")
    )
  ) +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  tm_title("d) LGS, k = 1000") +
  borders +
  zoom_borders

# Save maps
tmap_save(rue_pgpp_map,
          here("results", "figures", "rue_pgpp.png"),
          width = 12, height = 12, units = "cm", dpi = 250)

tmap_save(rpi_pgpp_map,
          here("results", "figures", "rpi_pgpp.png"),
          width = 12, height = 12, units = "cm", dpi = 250)

tmap_save(lgs_100_pgpp_map,
          here("results", "figures", "lgs_100_pgpp.png"),
          width = 12, height = 12, units = "cm", dpi = 250)

tmap_save(lgs_200_mean_map,
          here("results", "figures", "lgs_200_mean.png"),
          width = 12, height = 12, units = "cm", dpi = 250)

tmap_save(lgs_1000_pgpp_map,
          here("results", "figures", "lgs_1000_pgpp.png"),
          width = 12, height = 12, units = "cm", dpi = 250)

```
