==> analysis/7_results_temporal.Rmd <==
---
title: "Results"
output: html_notebook
author: Guy Lomax
date: 2024-01-12
editor_options: 
  chunk_output_type: inline
---

This notebook compares the results of the RPI methodology with those of RUE,
RESTREND and LGS.

```{r setup}

library(tidyverse)
library(terra)
library(sf)
library(here)
library(tmap)

```


```{r load}

# Country polygons
ke_tz <- st_read(here("data", "raw", "vector", "kenya_tanzania.geojson"))

# Potential and actual GPP values from RPI model
rpi_paths <- Sys.glob(here("data", "processed", "raster", "rpi", "rf_rpi_20*.tif"))

potential_gpp_all <- rpi_paths %>%
  map(function(path) rast(path) %>% subset("quantile_pred")) %>%
  rast()
names(potential_gpp_all) <- paste0("potential_", 2000:2018)

actual_gpp_all <- rpi_paths %>%
  map(function(path) rast(path) %>% subset("GPP")) %>%
  rast()
names(actual_gpp_all) <- paste0("actual_", 2000:2018)

mean_gpp_all <- rpi_paths %>%
  map(function(path) rast(path) %>% subset("mean_pred")) %>%
  rast()
names(mean_gpp_all) <- paste0("mean_pred_", 2000:2018)

# RPI layers and trend

rpi_all <- rpi_paths %>%
  map(function(path) rast(path) %>% subset("rpi")) %>%
  rast()
names(rpi_all) <- paste0("rpi_", 2000:2018)

rpi_ts_trend <- rast(here("data", "processed", "raster", "rpi", "rf_rpi_ts_trend.tif"))

# RUE layers and trend
rue_all <- rast(here("data", "processed", "raster", "rue", "rue_all.tif"))
rue_ts_trend <- rast(here("data", "processed", "raster", "rue", "rue_trend.tif"))

names(rue_all) <- paste0("rue_", 2000:2018)

# RESTREND residuals and trend
restrend <- rast(here("data", "processed", "raster", "restrend", "restrend_resids.tif"))
restrend_residuals <- restrend %>%
  subset(str_detect(names(restrend), "resid"))
restrend_results <- rast(here("data", "processed", "raster", "restrend", "restrend_results.tif"))

names(restrend_residuals) <- paste0("restrend_resids_", 2000:2018)

# LGS layers
lgs_annual_paths <- Sys.glob(here("data", "processed", "raster", "lgs", "lgs_annual_*.tif"))
lgs_potential_gpp_all <- lgs_annual_paths %>%
  map(function(path) rast(path) %>% subset("potential")) %>%
  rast()
names(lgs_potential_gpp_all) <- paste0("lgs_potential_", 2000:2018)

lgs_mean_gpp_all <- lgs_annual_paths %>%
  map(function(path) rast(path) %>% subset("mean")) %>%
  rast()

# Key covariates for comparison/exploration of results
dynamic_covariate_paths <- Sys.glob(here("data", "raw", "raster", "covariateMaps", "dynamicVars*.tif"))
ppt_all <- dynamic_covariate_paths %>%
  map(function(path) rast(path) %>% subset("precipitation")) %>%
  rast()
tMean_all <- dynamic_covariate_paths %>%
  map(function(path) rast(path) %>% subset("tMean")) %>%
  rast()
tree_cover <- rast(here("data", "raw", "raster", "covariateMaps", "staticVars.tif")) %>%
  subset("wriTCFrac") %>%
  crop(ppt_all)

years <- 2000:2018
names(ppt_all) <- paste0("precipitation_", years)
names(tMean_all) <- paste0("tMean_", years)

```


Now we calculate the R-squared and the weighted mean absolute error per pixel
for each method.

The weighted mean absolute error is the mean absolute difference between actual and
predicted GPP for each year in a pixel. Methods with the lowest pixel-wise MAE
explain the greatest amount of variance as a function of environmental
covariates. Note that the MAE for pixel-wise methods (RESTREND) are
likely to be lower, since the pixel-wise model only needs to explain variance
over time in each pixel, rather than also explaining spatial variance.

For the RUE and RPI methods, we are interested here only in ability to explain
the temporal variability in GPP, not the spatial variability. We therefore
compare the annual anomaly in GPP predicted by RPI (predicted GPP - mean
predicted GPP) to that observed for that pixel (observed GPP - mean observed GPP)
when calculating mean error and R-squared.

```{r pixel_wise_error}

# First, calculate the actual GPP anomaly (difference between annual GPP
# and mean GPP for that pixel)

actual_gpp_anomaly <- actual_gpp_all - mean(actual_gpp_all)

# RUE

rue_mean <- mean(actual_gpp_all, na.rm = T) / mean(ppt_all, na.rm = T)

rue_pred <- ppt_all * rue_mean

rue_pred_anomaly <- rue_pred - mean(rue_pred)

gpp_error_rue <- abs(rue_pred_anomaly - actual_gpp_anomaly)

mae_rue <- mean(gpp_error_rue)

# RESTREND

mae_restrend <- mean(abs(restrend_residuals))

# RPI

rpi_gpp_anomaly <- mean_gpp_all - mean(mean_gpp_all)

gpp_error_rpi <- rpi_gpp_anomaly - actual_gpp_anomaly

mae_rpi_anom <- mean(abs(gpp_error_rpi))

# Best method per pixel by MAE

best_mae <- which.min(c(mae_rue, mae_restrend, mae_rpi_anom))

```
Generate figures of mean absolute error per grid cell for each method
(Supplementary Figure S3).

```{r mae_figs}

indices <- c("RUE", "RESTREND", "RPI")

fill <- tm_shape(ke_tz) +
  tm_fill("grey95")

borders <- tm_shape(ke_tz) +
  tm_borders()

mae_raster_params <- tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(0, 1),
      values = "or_rd",
      outliers.trunc = c(TRUE, TRUE)
    ),
    col.legend = tm_legend(
      title = expression(
      atop("GPP MAE",
           (kg~C~m^-2~y^-1))
    ),
      reverse = TRUE,
      frame = FALSE
    )
  ) +
  tm_layout(asp = 3/4, frame = FALSE)

rue_mae_map <- fill +
  tm_shape(mae_rue * 0.001) +
  mae_raster_params +
  borders +
  tm_title("b) RUE")

restrend_mae_map <- fill +
  tm_shape(mae_restrend * 0.001) +
  mae_raster_params +
  borders +
  tm_title("c) RESTREND")

rpi_mae_map <- fill +
  tm_shape(mae_rpi_anom * 0.001) +
  mae_raster_params +
  borders +
  tm_title("a) RPI")

best_mae_map <- fill +
  tm_shape(best_mae) +
  tm_raster(
    col.scale = tm_scale_categorical(
      values = c("#C68039", "#8039C6", "#39C680"),
      levels = c(1,2,3),
      labels = c("RUE", "RESTREND", "RPI")
    ),
    col.legend = tm_legend(title = "",
                           frame = FALSE)
  ) +
  tm_layout(asp = 3/4, frame = FALSE) +
  tm_title("d) Best method") +
  borders

tmap_save(
  rue_mae_map,
  here("results", "figures", "rue_annual_mae_map.png"),
       width = 10, height = 10, units = "cm", dpi = 250
)

tmap_save(
  rpi_mae_map,
  here("results", "figures", "rpi_annual_mae_map.png"),
       width = 10, height = 10, units = "cm", dpi = 250
)

tmap_save(
  restrend_mae_map,
  here("results", "figures", "restrend_annual_mae_map.png"),
       width = 10, height = 10, units = "cm", dpi = 250
)

tmap_save(
  best_mae_map,
  here("results", "figures", "best_mae_map.png"),
       width = 10, height = 10, units = "cm", dpi = 250
)

# Histogram

combined_mae <- c(mae_rue, mae_restrend, mae_rpi_anom) * 0.001
names(combined_mae) <- indices

mae_df <- as.data.frame(combined_mae, na.rm = T) %>%
  pivot_longer(cols = all_of(indices),
               names_to = "method", values_to = "mae") %>%
  mutate(method = ordered(method, levels = c("RUE", "RESTREND", "RPI")))
  

mae_hist <- ggplot(mae_df, aes(x = mae, fill = method)) +
  geom_density(position = "identity", alpha = 0.5, colour = "grey30") +
  theme_minimal() +
  scale_fill_manual(values = c("#C68039", "#8039C6", "#39C680")) +
  theme(title = element_text(size = 16),
        axis.title = element_text(size = 20),
        axis.text = element_text(size = 16)) +
  xlim(0, 1.5) +
  labs(x = expression(MAE~(kg~C~m^-2~y^-1)),
       y = "Density",
       fill = "Method",
       title = "e) Distribution of pixel-level MAE over study region")
    
ggsave(
  here("results", "figures", "mae_hist.png"),
  mae_hist,
  width = 24, height = 12, units = "cm", dpi = 250
)




```

Calculate pixel-wise r-squared using pixel-wise total sum of squares and residual
sum of squares.


```{r pixel_wise_r_sq}

# Total variance in GPP
actual_gpp_mean <- mean(actual_gpp_all)

actual_gpp_anomaly <- actual_gpp_all - actual_gpp_mean

# Total sum of squares
actual_gpp_tss <- sum(actual_gpp_anomaly ^ 2)

# RUE

rue_rss <- sum(gpp_error_rue ^ 2)

rue_rsq <- 1 - (rue_rss / actual_gpp_tss)

rue_rsq_lt0 <- rue_rsq < 0
rue_rsq_lt0 <- mask(rue_rsq_lt0, rue_rsq_lt0, maskvalues = c(0,NA))

# RESTREND

restrend_rss <- sum(restrend_residuals ^ 2)

restrend_rsq <- 1 - (restrend_rss / actual_gpp_tss)

# RPI

rpi_rss <- sum(gpp_error_rpi ^ 2)

rpi_rsq <- 1 - (rpi_rss / actual_gpp_tss)

rpi_rsq_lt0 <- rpi_rsq < 0
rpi_rsq_lt0 <- mask(rpi_rsq_lt0, rpi_rsq_lt0, maskvalues = c(0, NA))

# Raster of best performing strategy

best_rsq <- which.max(c(rue_rsq, restrend_rsq, rpi_rsq))

```

Generate figures of R-squred per grid cell for each method (Figure 5).

```{r r_sq_figs}
indices <- c("RUE", "RESTREND", "RPI")

fill <- tm_shape(ke_tz) + 
  tm_fill("grey95")

borders <- tm_shape(ke_tz) +
  tm_borders()

rsq_raster_params <- tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(0, 0.75),
      # midpoint = 0,
      values = "yl_gn",
      outliers.trunc = c(TRUE, TRUE)
    ),
    col.legend = tm_legend(
      title = expression(R^2),
      reverse = TRUE,
      frame = FALSE
    )
  ) 


rue_rsq_map <- fill +
  tm_shape(rue_rsq) +
  rsq_raster_params +
  tm_shape(rue_rsq_lt0) +
  tm_raster(col.scale = tm_scale_categorical(values = "lightsalmon", labels = expression(Negative~R^2)),
            col.legend = tm_legend(title = "")) +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  borders +
  tm_title("b) RUE")

rpi_rsq_map <- fill +
  tm_shape(rpi_rsq) +
  rsq_raster_params +
  tm_shape(rpi_rsq_lt0) +
  tm_raster(col.scale = tm_scale_categorical(values = "lightsalmon", labels = expression(Negative~R^2)),
            col.legend = tm_legend(title = "")) +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  borders +
  tm_title("a) RPI")

restrend_rsq_map <- fill +
  tm_shape(restrend_rsq) +
  rsq_raster_params +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  borders +
  tm_title("c) RESTREND")

# rue_rsq_map
# restrend_rsq_map
# rpi_rsq_map

tmap_save(
  rue_rsq_map,
  here("results", "figures", "rue_annual_rsq_map.png"),
  width = 10, height = 10, units = "cm", dpi = 250
)

tmap_save(
  rpi_rsq_map,
  here("results", "figures", "rpi_annual_rsq_map.png"),
  width = 10, height = 10, units = "cm", dpi = 250
)

tmap_save(
  restrend_rsq_map,
  here("results", "figures", "restrend_annual_rsq_map.png"),
  width = 10, height = 10, units = "cm", dpi = 250
)

# Map of best performing strategy

best_method_map <- fill +
  tm_shape(best_rsq) +
  tm_raster(
    col.scale = tm_scale_categorical(
      values = c("#C68039", "#8039C6", "#39C680"),
      levels = c(1,2,3),
      labels = c("RUE", "RESTREND", "RPI")
    ),
    col.legend = tm_legend(
      title = "",
      frame = FALSE
    ),
    # col_alpha = 0.8
  ) +
  tm_title("d) Best method") +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  borders

tmap_save(
  best_method_map,
  here("results", "figures", "best_method_rsq.png"),
  width = 10, height = 10, units = "cm", dpi = 250
)

# Histogram
combined_rsq <- c(rue_rsq, restrend_rsq, rpi_rsq)

names(combined_rsq) <- indices

rsq_df <- as.data.frame(combined_rsq, na.rm = T) %>%
  pivot_longer(cols = all_of(indices),
               names_to = "method", values_to = "rsq") %>%
  mutate(method = ordered(method, levels = c("RUE", "RESTREND", "RPI")))

rsq_hist <- ggplot(rsq_df, aes(x = rsq, fill = method)) +
  geom_density(position = "identity", alpha = 0.6, colour = "grey30") +
  geom_vline(xintercept = 0) +
  theme_minimal() +
  scale_fill_manual(values = c("#C68039", "#8039C6", "#39C680")) +
  theme(title = element_text(size = 16),
        axis.title = element_text(size = 20),
        axis.text = element_text(size = 16)) +
  xlim(-2, 1) +
  labs(x = expression(R^2),
       y = "Density",
       fill = "Method",
       title = expression("e) Distribution of pixel-level"~R^2~"values over study region"))

ggsave(
  here("results", "figures", "rsq_hist.png"),
  rsq_hist,
  width = 24, height = 12, units = "cm", dpi = 250
)
  


```

Calculate and plot pixel-wise correlation between indices and annual
precipitation (Supplementary Figure S4).

```{r ppt_correlation_map}

raster_to_tidy_df <- function(raster, xy = FALSE) {
  df <- as.data.frame(raster, xy = xy, cells = TRUE)
  
  var_name_year <- names(raster)[1]
  var_name_length <- nchar(var_name_year) - 5
  var_name <- substring(var_name_year, 1, var_name_length)
  
  tidy_df <- df %>%
    pivot_longer(cols = starts_with(var_name),
                 names_to = "year",
                 values_to = var_name) %>%
    mutate(year = as.numeric(substring(year, var_name_length + 2)))
  
  tidy_df
}

ppt_df <- raster_to_tidy_df(ppt_all, xy = TRUE)
rue_df <- raster_to_tidy_df(rue_all)
restrend_df <- raster_to_tidy_df(restrend_residuals)
rpi_df <- raster_to_tidy_df(rpi_all)

combined_df_all <- ppt_df %>%
  left_join(rue_df) %>%
  left_join(restrend_df) %>%
  left_join(rpi_df)

ppt_cor_all <- combined_df_all %>%
  pivot_longer(cols = c(rue, restrend_resids, rpi), names_to = "method") %>%
  group_by(cell, method) %>%
  summarise(ppt_cor = cor(precipitation, value),
            x = first(x),
            y = first(y)) %>%
  ungroup()

# Convert back to correlation raster layers

ppt_cor_rast <- ppt_cor_all %>%
  pivot_wider(id_cols = c(cell, x, y),
              values_from = ppt_cor,
              names_from = method) %>%
  select(x, y, rue, restrend_resids, rpi) %>%
  rast(crs = crs(rpi_all), extent = ext(rpi_all))

best_cor <- which.min(abs(ppt_cor_rast))

# Save maps

fill <- tm_shape(ke_tz) + 
  tm_fill("grey90")

borders <- tm_shape(ke_tz) +
  tm_borders()

ppt_cor_params <- tm_raster(
  col.scale = tm_scale_continuous(
    limits = c(-1, 1),
    midpoint = 0,
    values = "PuOr"
  ),
  col.legend = tm_legend(
    title = "Correlation",
    frame = FALSE,
    reverse = TRUE
  )
) + tm_layout(
  asp = 3/4,
  frame = FALSE
)

rpi_ppt_cor <- fill +
  tm_shape(ppt_cor_rast$rpi) +
  ppt_cor_params +
  borders +
  tm_title("a) RPI")

rue_ppt_cor <- fill +
  tm_shape(ppt_cor_rast$rue) +
  ppt_cor_params +
  borders +
  tm_title("b) RUE")

restrend_ppt_cor <- fill +
  tm_shape(ppt_cor_rast$restrend_resids) +
  ppt_cor_params +
  borders +
  tm_title("c) RESTREND")

best_ppt_cor <- fill +
  tm_shape(best_cor) +
  tm_raster(
    col.scale = tm_scale_categorical(
      values = c("#C68039", "#8039C6", "#39C680"),
      levels = c(1,2,3),
      labels = c("RUE", "RESTREND", "RPI")
    ),
    col.legend = tm_legend(
      title = "",
      frame = FALSE
    )
  ) +
  tm_layout(asp = 3/4,
            frame = FALSE) +
  borders +
  tm_title("d) Best method")

tmap_save(
  rpi_ppt_cor,
  here("results", "figures", "rpi_ppt_cor.png"),
  width = 10, height = 10, units = "cm", dpi = 250
)

tmap_save(
  rue_ppt_cor,
  here("results", "figures", "rue_ppt_cor.png"),
  width = 10, height = 10, units = "cm", dpi = 250
)

tmap_save(
  restrend_ppt_cor,
  here("results", "figures", "restrend_ppt_cor.png"),
  width = 10, height = 10, units = "cm", dpi = 250
)

tmap_save(
  best_ppt_cor,
  here("results", "figures", "best_method_ppt_cor.png"),
  width = 10, height = 10, units = "cm", dpi = 250
)

# Density plot of correlations for each method

ppt_cor_hist <- ppt_cor_all %>%
  mutate(method = factor(method, levels = c("rue", "restrend_resids", "rpi"))) %>%
  ggplot(aes(x = ppt_cor, fill = method)) +
  geom_density(position = "identity", alpha = 0.6, colour = "grey30") +
  geom_vline(xintercept = 0) +
  theme_minimal() +
  scale_fill_manual(values = c("#C68039", "#8039C6", "#39C680"),
                    labels = c("RUE", "RESTREND", "RPI")) +
  theme(title = element_text(size = 16),
        axis.title = element_text(size = 20),
        axis.text = element_text(size = 16)) +
  coord_cartesian(xlim = c(-1, 1), ylim = c(0, 2)) +
  labs(x = "R",
       y = "Density",
       fill = "Method",
       title = "e) Correlation with annual precipitation per grid cell")

ggsave(
  here("results", "figures", "ppt_cor_hist.png"),
  ppt_cor_hist,
  width = 24, height = 12, units = "cm", dpi = 250
)


# Area % where each method is best
area_best <- tibble(value = 1:3, method = c("RUE", "RESTREND", "RPI")) %>%
  mutate(area_best = map_vec(value, function(val) {
    sum(
      (values(best_cor) == val) * values(cellSize(best_cor, unit = "km")),
      na.rm = TRUE)
    }
  )) %>%
  mutate(area_frac = area_best / sum(area_best))

```


Scatter plots of temporal trends from the three time-series methods
(Supplementary Figure S5).

```{r trend_scatter_plots}

options(scipen = 999)

set.seed(121)
rpi_trend_sample <- spatSample(rpi_ts_trend$slope, 4000, na.rm = TRUE, exhaustive = TRUE, as.points = TRUE)

rue_trend_sample <- extract(rue_ts_trend$rue_slope, rpi_trend_sample)
restrend_sample <- extract(restrend_results$resid_slope, rpi_trend_sample)
mean_ppt_sample <- ppt_all %>%
  mean() %>%
  extract(rpi_trend_sample)

trend_sample_all <- rpi_trend_sample %>%
  as.data.frame(geom = "XY") %>%
  bind_cols(rue_trend_sample, restrend_sample, mean_ppt_sample) %>%
  select(-starts_with("ID")) %>%
  rename(precipitation = mean,
         rpi_slope = slope,
         restrend_slope = resid_slope) %>%
  mutate(ppt_bin = cut_width(precipitation, 200, boundary = 0))

bins <- levels(trend_sample_all$ppt_bin)
ppt_labels <- c(
  expression("<200"~mm~yr^-1),
  expression("200-400"~mm~yr^-1),
  expression("400-600"~mm~yr^-1),
  expression("600-800"~mm~yr^-1),
  expression("800-1000"~mm~yr^-1),
  expression("1000-1200"~mm~yr^-1))
# names(ppt_labels) <- as.character(bins[1:6])

trend_sample_subset <- filter(trend_sample_all, ppt_bin %in% bins[1:6])

trend_sample_subset$ppt_bin <- factor(trend_sample_subset$ppt_bin, labels = ppt_labels)

rpi_rue_plot <- ggplot(trend_sample_subset, aes(x = rpi_slope, y = rue_slope)) +
  geom_point(size = 0.2) +
  xlim(-0.03, 0.03) +
  ylim(-0.2, 0.2) +
  geom_hline(yintercept = 0) +
  geom_vline(xintercept = 0) +
  facet_wrap(~ppt_bin, labeller = label_parsed) +
  theme_bw() +
  labs(x = expression("RPI trend"~(yr^-1)), y = expression("RUE trend"~(g~C~m^-2~yr^-2~mm^-1))) +
  theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 14),
        strip.text = element_text(size = 14))

rpi_restrend_plot <- ggplot(trend_sample_subset, aes(x = rpi_slope, y = restrend_slope)) +
  geom_point(size = 0.2) +
  geom_smooth(method = "lm", se = FALSE) +
  xlim(-0.03, 0.03) +
  # ylim(-0.2, 0.2) +
  geom_hline(yintercept = 0) +
  geom_vline(xintercept = 0) +
  facet_wrap(~ppt_bin, labeller = label_parsed) +
  theme_bw() +
  labs(x = expression("RPI trend"~(yr^-1)), y = expression("RESTREND trend"~(g~C~m^-2~yr^-2))) +
  theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 14),
        strip.text = element_text(size = 14))

ggsave(here("results", "figures", "rpi_vs_rue_facets.png"), rpi_rue_plot,
       width = 24, height = 16, units = "cm", dpi = 200)
ggsave(here("results", "figures", "rpi_restrend_plot.png"), rpi_restrend_plot,
       width = 24, height = 16, units = "cm", dpi = 200)

```

Annual facets of RPI values (Figure 3).

```{r facet_maps_rpi}

names(rpi_all) <- years

fill <- tm_shape(ke_tz) + 
  tm_fill("grey95")

borders <- tm_shape(ke_tz) +
  tm_borders()

rpi_annual_map <- fill +
  tm_shape(rpi_all[[2:19]]) +
  tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(0.2, 1),
      values = "viridis",
      outliers.trunc = c(TRUE, TRUE)
    ),
    col.legend = tm_legend(show = FALSE)
  ) +
  borders +
  tm_layout(asp = 3/4, frame = FALSE,
            panel.label.bg.color = "white",
            panel.label.size = 10,
            panel.label.frame = FALSE) +
  tm_facets_wrap(nrow = 3, ncol = 6)

rpi_map_legend <- tm_shape(rpi_all[[2]]) +
  tm_raster(
    col.scale = tm_scale_continuous(
      limits = c(0.2, 1),
      values = "viridis",
      values.range = c(0,1),
      outliers.trunc = c(TRUE, TRUE)
    ),
    col.legend = tm_legend(
      title = "RPI",
      reverse = TRUE,
      frame = FALSE
    ),
  ) +
  tm_layout(legend.only = TRUE)

tmap_save(
  rpi_annual_map,
  here("results", "figures", "rpi_annual_map.png"),
  width = 30, height = 30, units = "cm", dpi = 250
)

tmap_save(
  rpi_map_legend,
  here("results", "figures", "rpi_map_legend.png"),
  width = 3, height = 6, units = "cm", dpi = 250
)

```

Calculate RPI autocorrelation to test whether inter-annual effects are important.

```{r rpi_autocorrelation}

rpi_detrended <- app(rpi_all, function(x) {
  
  if(any(is.na(x))) {
    rep(NA, 19)
  } else {
    years <- 1:19
    
    trend <- lm(x ~ years)
    
    residuals(trend)
    
  }
})

calc_ac1 <- function(x) {
  if(any(is.na(x))) {
    NA
  } else {
    ac_all <- acf(x, lag.max = 2, plot = FALSE)
    
    ac_all$acf[2]
  }
}

rpi_ac1 <- app(rpi_all, calc_ac1)
rpi_ac1_detrended <- app(rpi_detrended, calc_ac1)

rpi_ac1_detrended

```
