==> analysis/8_rpi_pattern_maps.Rmd <==
---
title: "RPI Pattern Maps"
output: html_notebook
author: Guy Lomax
date: 2024-06-18
---

This notebook allows plotting of RPI spatial and temporal patterns in the
study region and specific regions of interest.

```{r setup, include = FALSE}

library(tidyverse)
library(terra)
library(sf)
library(tmap)
library(here)
library(units)

```


Load data

```{r load}

# Country polygons
ke_tz <- st_read(here("data", "raw", "vector", "kenya_tanzania.geojson"))

# Potential and actual GPP values from RPI model
rpi_paths <- Sys.glob(here("data", "processed", "raster", "rpi", "rf_rpi_20*.tif"))

potential_gpp_all <- rpi_paths %>%
  map(function(path) rast(path) %>% subset("quantile_pred")) %>%
  rast()
names(potential_gpp_all) <- paste0("potential_", 2000:2018)

actual_gpp_all <- rpi_paths %>%
  map(function(path) rast(path) %>% subset("GPP")) %>%
  rast()
names(actual_gpp_all) <- paste0("actual_", 2000:2018)

mean_gpp_all <- rpi_paths %>%
  map(function(path) rast(path) %>% subset("mean_pred")) %>%
  rast()
names(mean_gpp_all) <- paste0("mean_pred_", 2000:2018)

# RPI layers and trend

rpi_all <- rpi_paths %>%
  map(function(path) rast(path) %>% subset("rpi")) %>%
  rast()
names(rpi_all) <- paste0("rpi_", 2000:2018)

rpi_ts_trend <- rast(here("data", "processed", "raster", "rpi", "rf_rpi_ts_trend.tif"))
rpi_kendall <- rast(here("data", "processed", "raster", "rpi", "rf_rpi_kendall_tau_rast.tif"))

# Topography

elev <- rast(here("data", "raw", "raster", "covariateMaps", "dem.tif"))

```


Quantify RPI levels and trends using quantiles

```{r rpi_patterns}

rpi_mean <- mean(rpi_all, na.rm = TRUE)

rpi_ts_quantiles <- quantile(values(rpi_ts_trend$slope), seq(0, 1, 0.2), na.rm = TRUE)
rpi_mean_quantiles <- quantile(values(rpi_mean), seq(0, 1, 0.2), na.rm = TRUE)

# rpi_ts_sig_trend <- ((rpi_ts_trend$slope > 0) - (rpi_ts_trend$slope < 0)) * (rpi_ts_trend$p_value < 0.05)

rpi_ts_classes <- tibble(
  lower = rpi_ts_quantiles[1:5],
  upper = lead(rpi_ts_quantiles)[1:5],
  class = 1:5
)

rpi_mean_classes <- tibble(
  lower = rpi_mean_quantiles[1:5],
  upper = lead(rpi_mean_quantiles)[1:5],
  class = 1:5
)

rpi_ts_classes_abs <- tibble(
  lower = seq(-0.1, 0.1, 0.04)[1:5],
  upper = lead(seq(-0.1, 0.1, 0.04))[1:5],
  class = 1:5
)

rpi_ts_reclass_abs <- classify(rpi_ts_trend$slope, rpi_ts_classes_abs)

rpi_ts_reclass <- classify(rpi_ts_trend$slope, rpi_ts_classes)
rpi_mean_reclass <- classify(rpi_mean, rpi_mean_classes)

# Create combined quantile index
rpi_combined_index <- rpi_mean_reclass * 10 + rpi_ts_reclass

# Visualise according to the following classes:
# 1. Persistent low RPI - mean RPI in bottom quintile and trend in bottom 4 quintiles
#     Values: 11, 12, 13, 14
# 2. Persistent high RPI - mean RPI in top quintile and trend in top 4 quintiles
#     Values: 52, 53, 54, 55
# 3. Strong decrease - mean RPI in middle 3 quintiles and trend in bottom quintile
#     Values: 21, 31, 41, 51
# 3. Strong increase - mean RPI in middle 3 quintiles and trend in top quintile
#     Values: 15, 25, 35, 45
# 5. Neutral - all others

rpi_combined_class <- tibble(
  from = c(
    11, 12, 13, 14,
    52, 53, 54, 55,
    21, 31, 41, 51,
    15, 25, 35, 45
  ),
  to = c(
    1, 1, 1, 1,
    2, 2, 2, 2,
    3, 3, 3, 3,
    4, 4, 4, 4
  )
)

rpi_combined_reclass <- classify(rpi_combined_index, rpi_combined_class, others = 5)

### Map plots

## Create new bbox to allow space for legend

map_bbox <- st_bbox(ke_tz)

xrange <- map_bbox$xmax - map_bbox$xmin
yrange <- map_bbox$ymax - map_bbox$ymin

map_bbox[1] <- map_bbox[1] - 0.25 * xrange

## Base map elements

fill <- tm_shape(ke_tz, bbox = map_bbox) + 
  tm_fill("grey95")

borders <- tm_shape(ke_tz) +
  tm_borders(lwd = 2)

## Colour scales

mean_scale <- tm_scale_continuous(
  limits = c(0.3, 1),
  values = "-viridis",
  outliers.trunc = c(TRUE, TRUE)
)

trend_scale <- tm_scale_continuous(
  limits = c(-0.025, 0.025),
  midpoint = 0,
  values = "pu_gn_div",
  outliers.trunc = c(TRUE, TRUE)
)

rpi_class_palette <- c("#A40000", "#16317D", "#FFCD12", "#007E2F", "#CCCCCC")
rpi_class_labels <- c("Persistent low", "Persistent high", "Strong decrease", "Strong increase", "Other")

class_scale <- tm_scale_categorical(
  values = rpi_class_palette,
  levels = c(1,2,3,4,5),
  labels = rpi_class_labels
)

# writeRaster(rpi_combined_reclass, here("results", "rpi_pattern_classes.tif"), overwrite = TRUE)

## Inset map boxes

# Serengeti-Mara
bb1 <- st_bbox(c(xmin = 34, xmax = 36, ymin = -3.5, ymax = -1.5), crs = "EPSG:4326") %>% st_as_sfc() %>% st_as_sf()

# # Tarangire and Arusha
# bb2 <- st_bbox(c(xmin = 35, xmax = 38, ymin = -5.25, ymax = -2.25), crs = "EPSG:4326") %>% st_as_sfc() %>% st_as_sf()
# 
# # Tana river
# bb3 <- st_bbox(c(xmin = 38, xmax = 41, ymin = -1.75, ymax = 1.25), crs = "EPSG:4326") %>% st_as_sfc() %>% st_as_sf()

# Northern Rangelands
bb4 <- st_bbox(c(xmin = 36.5, xmax = 38.5, ymin = 0, ymax = 2), crs = "EPSG:4326") %>% st_as_sfc() %>% st_as_sf()



bb_list <- list(bb1, bb4)
bb_sf <- bind_rows(bb_list) %>% mutate(id = 1:nrow(.))
bb_colour_list <- c("orange2", "green2")

bb_map <- tm_shape(bb_sf) + 
  tm_borders(
    col = "id", 
    col.scale = tm_scale_discrete(values = bb_colour_list), 
    col.legend = tm_legend(show = FALSE),
    lwd = 2.5)

## Large-scale maps

rpi_mean_map <- fill +
  tm_shape(rpi_mean) +
  tm_raster(
    col.scale = mean_scale,
    col.legend = tm_legend(
      title = "RPI Mean",
      frame = FALSE,
      # orientation = "landscape",
      reverse = TRUE,
      position = c("left", "top"),
      height = 15
    )
  ) +
  bb_map +
  tm_layout(frame = FALSE, asp = 4/3) +
  # tm_title("a)") +
  borders

rpi_trend_map <- fill +
  tm_shape(rpi_ts_trend$slope) +
  tm_raster(
    col.scale = trend_scale,
    col.legend = tm_legend(
      title = "RPI Trend Slope",
      frame = FALSE,
      # orientation = "landscape",
      reverse = TRUE,
      position = c("left", "top"),
      height = 15
    )
  ) +
  bb_map +
  tm_layout(frame = FALSE, asp = 4/3) +
  # tm_title("d)") +
  borders

rpi_class_map <- fill +
  tm_shape(rpi_combined_reclass) +
  tm_raster(
    col.scale = class_scale,
    col.legend = tm_legend(
      # show = FALSE,
      title = "RPI Patterns",
      frame = FALSE,
      # orientation = "landscape",
      position = c("left", "top"),
      height = 12
    )
  ) +
  bb_map +
  tm_layout(frame = FALSE, asp = 4/3) +
  # tm_title("g)") +
  borders

tmap_save(rpi_mean_map, here("results", "figures", "rpi_mean_map.png"),
          width = 12, height = 12, units = "cm", dpi = 300)
tmap_save(rpi_trend_map, here("results", "figures", "rpi_trend_map.png"),
          width = 12, height = 12, units = "cm", dpi = 300)
tmap_save(rpi_class_map, here("results", "figures", "rpi_class_map.png"),
          width = 12, height = 12, units = "cm", dpi = 300)

## Zoom-ins

# Create cropped versions of rasters to each bbox (otherwise creates issues downsampling)

rast_combined <- c(rpi_mean, rpi_ts_trend$slope, rpi_combined_reclass)
names(rast_combined) <- c("mean", "trend", "class")

rast1 <- crop(rast_combined, bb1)
# rast2 <- crop(rast_combined, bb2)
# rast3 <- crop(rast_combined, bb3)
rast4 <- crop(rast_combined, bb4)

map_list <- list(list(rpi_mean_map, rpi_trend_map, rpi_class_map),
                 list(length = 3), 
                 list(length = 3))

for (i in 1:length(bb_list)) {
  bb <- bb_list[[i]]
  
  bb_rast <- crop(rast_combined, bb)
  
  fill_bb <- tm_shape(ke_tz, bbox = bb) + tm_fill("white")
  bb_border <- tm_shape(bb) + tm_borders(col = bb_colour_list[i], lwd = 2)
  
  rpi_mean_bb <- fill_bb +
    tm_shape(bb_rast$mean) +  tm_raster(
      col.scale = mean_scale,
      col.legend = tm_legend(show = FALSE),
      col.alpha = 0.5
    ) +
    tm_layout(asp = 3/4,
            frame = FALSE) + 
    # tm_title(paste0(letters[1 + i], ")")) +
    bb_border
  
  rpi_trend_bb <- fill_bb +
    tm_shape(bb_rast$trend) +  tm_raster(
      col.scale = trend_scale,
      col.legend = tm_legend(show = FALSE),
      col.alpha = 0.5
    ) +
    tm_layout(asp = 3/4,
            frame = FALSE) +
    # tm_title(paste0(letters[4 + i], ")")) +
    bb_border
  
  rpi_class_bb <- fill_bb +
    tm_shape(bb_rast$class) +  tm_raster(
      col.scale = class_scale,
      col.legend = tm_legend(show = FALSE),
      col.alpha = 0.5
    ) +
    tm_layout(asp = 3/4,
            frame = FALSE) +
    # tm_title(paste0(letters[7 + i], ")")) +
    bb_border
  
  map_list[[i+1]] <- list(rpi_mean_bb, rpi_trend_bb, rpi_class_bb)
  
}

tmap_list <- list_flatten(list_transpose(map_list))

multi_panel_map <- tmap_arrange(tmap_list, ncol = 3, nrow = 3)

tmap_save(multi_panel_map, here("results", "figures", "multi_panel_class_map.jpg"),
          width = 26, height = 32, dpi = 300, units = "cm")

## High topography area map

# High topography around Ngorongoro
bb_topo <- st_bbox(c(xmin = 35.2, xmax = 36.2, ymin = -3.8, ymax = -2.8), crs = "EPSG:4326") %>% st_as_sfc() %>% st_as_sf()

rpi_mean_map_with_topo_bbox <- fill +
  tm_shape(rpi_mean) +
  tm_raster(
    col.scale = mean_scale,
    col.legend = tm_legend(
      title = "RPI Mean",
      frame = FALSE,
      # orientation = "landscape",
      reverse = TRUE,
      position = c("left", "top"),
      height = 15
    )
  ) +
  tm_shape(bb_topo) + tm_borders(col = "black", lwd = 2) +
  tm_layout(frame = FALSE) +
  borders

tmap_save(rpi_mean_map_with_topo_bbox, here("results", "figures", "rpi_mean_map_with_topo_bbox.png"),
          width = 10, height = 10, units = "cm", dpi = 300)

rpi_mean_topo <- crop(rpi_mean, bb_topo)

rpi_mean_topo_map <- tm_shape(ke_tz, bbox = bb_topo) + tm_fill(fill = "grey95") +
  tm_shape(rpi_mean_topo) + 
  tm_raster(
      col.scale = mean_scale,
      col.legend = tm_legend(show = FALSE),
      col.alpha = 0.5
    ) +
    tm_layout(frame = FALSE) +
  tm_shape(bb_topo) +
  tm_borders(col = "black", lwd = 2)

tmap_save(rpi_mean_topo_map, here("results", "figures", "rpi_mean_topo_map.png"),
          width = 10, height = 10, units = "cm", dpi = 300)

elev_bbox <- crop(elev, bb_topo)
elev_s_a <- terrain(elev_bbox, v = c("slope", "aspect"), unit = "radians")
hillshade <- shade(elev_s_a$slope, elev_s_a$aspect, angle = 315)

elev_topo_map <- tm_shape(ke_tz, bbox = bb_topo) + tm_fill(fill = "grey95") +
  tm_shape(elev_bbox) + 
  tm_raster(
      col.scale = tm_scale_continuous(values = "rainbow_bu_pk"),
      col.legend = tm_legend(position = c("left", "bottom"), reverse = TRUE, bg.color = "white", title = "Elevation (m)"),
      col.alpha = 0.5
    ) +
    tm_layout(frame = FALSE) +
  tm_shape(bb_topo) +
  tm_borders(col = "black", lwd = 2)

hillshade_map <- tm_shape(ke_tz, bbox = bb_topo) + tm_fill(fill = "grey95") +
  tm_shape(hillshade) + 
  tm_raster(
      col.scale = tm_scale_continuous(values = "-greys"),
      col.legend = tm_legend(show = FALSE),
      col.alpha = 0.5
    ) +
    tm_layout(frame = FALSE) +
  tm_shape(bb_topo) +
  tm_borders(col = "black", lwd = 2)

tmap_save(elev_topo_map, here("results", "figures", "elev_topo_map.png"),
          width = 12, height = 12, units = "cm", dpi = 300)
tmap_save(hillshade_map, here("results", "figures", "hillshade_topo_map.png"),
          width = 10, height = 10, units = "cm", dpi = 300)

```
